Fix for NPE occurring when HubProjectParameters not specified
When no HubProjectParameters are specified (ie, not providing
project_name and/or code_location at bind time), a NPE is thrown.
*Changed HubProjectParameters parameter in
BindingInstanceService.create() to Optional<HubProjectParameters>
*Changed call to create from BindingInstanceRestServer to pass
Optional<> and not default to null
*Changed BindingInstance.create() to access Project Name and Code
Location only if HubProjectParameters is present. If null is passed in
call to create, replace with Optional.empty().
*Added unit tests for BindingInstance.create()