Use configuration properties
Made application.yml the central location of all configuration items for
a particular project. The project loads from the application.yml into
classes which are then passed and used.
The application.yml used variable references to obtain values set in the
environment of the container so they can be configured at runtime.